+++ src/app/(content)/sermons/_component/SermonDetailPage/SermonMetaActions.tsx
@@ -1,6 +1,5 @@
 'use client';
 
-import { useEffect, useState } from 'react';
 import clsx from 'clsx';
 import {
   IoHeart,
@@ -11,11 +10,9 @@ import {
   IoShareSocialOutline
 } from 'react-icons/io5';
 import { BottomSheet } from '@/components/ui';
-import useKakaoShare from '@/hooks/useKakaoShare';
-import { useToastStore } from '@/store/toast.store';
+import { useSermonBookmark } from './useSermonBookmark';
+import { useSermonShare } from './useSermonShare';
 import styles from './SermonDetailPage.module.scss';
-
-const BOOKMARK_STORAGE_KEY = 'bookmarked-sermons';
 
 type Props = {
   sermonId: string;
@@ -24,81 +21,21 @@ type Props = {
   shareImageUrl?: string;
 };
 
-function loadBookmarkedSermonIds(): string[] {
-  if (typeof window === 'undefined') return [];
-  try {
-    const raw = window.localStorage.getItem(BOOKMARK_STORAGE_KEY);
-    const parsed = raw ? JSON.parse(raw) : [];
-    return Array.isArray(parsed)
-      ? parsed.filter((value): value is string => typeof value === 'string')
-      : [];
-  } catch {
-    return [];
-  }
-}
-
-function saveBookmarkedSermonIds(ids: string[]) {
-  try {
-    window.localStorage.setItem(BOOKMARK_STORAGE_KEY, JSON.stringify(ids));
-  } catch {
-    /* localStorage 접근 실패 시 무시 */
-  }
-}
-
 export default function SermonMetaActions({
   sermonId,
   title,
   description,
   shareImageUrl
 }: Props) {
-  const { info, error } = useToastStore();
-  const { share: shareToKakao } = useKakaoShare();
-  const [shareOpen, setShareOpen] = useState(false);
-  const [bookmarked, setBookmarked] = useState(false);
-
-  useEffect(() => {
-    setBookmarked(loadBookmarkedSermonIds().includes(sermonId));
-  }, [sermonId]);
-
-  const handleCopyLink = async () => {
-    try {
-      await navigator.clipboard.writeText(window.location.href);
-      info('링크가 복사되었습니다');
-    } catch {
-      error('링크 복사에 실패했습니다');
-    }
-    setShareOpen(false);
-  };
-
-  const handleKakaoShare = () => {
-    shareToKakao({ title, description, imageUrl: shareImageUrl });
-    setShareOpen(false);
-  };
-
-  const handleFacebookShare = () => {
-    const shareUrl = `https://www.facebook.com/sharer/sharer.php?u=${encodeURIComponent(
-      window.location.href
-    )}`;
-    window.open(shareUrl, '_blank', 'noopener,noreferrer,width=600,height=500');
-    setShareOpen(false);
-  };
-
-  const handleEmailShare = () => {
-    const subject = encodeURIComponent(title);
-    const body = encodeURIComponent(window.location.href);
-    window.location.href = `mailto:?subject=${subject}&body=${body}`;
-    setShareOpen(false);
-  };
-
-  const handleToggleBookmark = () => {
-    const current = loadBookmarkedSermonIds();
-    const next = bookmarked
-      ? current.filter((id) => id !== sermonId)
-      : [...current, sermonId];
-    saveBookmarkedSermonIds(next);
-    setBookmarked(!bookmarked);
-    info(bookmarked ? '저장이 해제되었습니다' : '저장되었습니다');
-  };
+  const { bookmarked, handleToggleBookmark } = useSermonBookmark(sermonId);
+  const {
+    shareOpen,
+    setShareOpen,
+    handleCopyLink,
+    handleKakaoShare,
+    handleFacebookShare,
+    handleEmailShare
+  } = useSermonShare({ title, description, shareImageUrl });
 
   return (
     <div className={styles.meta_actions}>
